Create workflow
This MVP is built with bash becuase I said to myself "I'll write a quick bash
script that uses curl and jq...". Several days later I have this
awesome/terrible bash script :), but it works™.

My future plan is to rebuild this workflow using ruby or go so that I can get
rid of the jq dependency and make it easier to add tests.