Improve matching
This adds the following improvements to matching:

1. ignores case
2. matches first letter of words split by spaces (ex. "ma" matches "my app app")
3. matches first letter of words split by hyphens (ex. "pds" matches
   "production-database-server")

As part of this, I also moved xml "rendering" out of `jq` and into a `for`
loop. This should be easier to read and maintain.

Note: improvements 2 and 3 are not quite where I would like them. For example,
"my awesome-app" gets tokenized as "ma" and "aa". I would prefer that to be
"maa".  Also, other characters are not filtered, so "my (awesome) app" gets
tokenized as "m(a". I may look at addressing those issues in the future, but I
also may have reached the limit of what I can reasonably do with curl + jq :).