Fix check for jq during install
The conditional that checks if `jq` is installed was based on a previous version of the install script where the executable was named `jq-osx-amd64`. I later updated the `wget` call to save the file as simply `jq`, but forgot to update the conditional. This fix will prevent `jq` from getting downloaded every time the install script runs.